+++ docs/mico-core/java/io/github/ust/mico/core/web/ApplicationController.rst
@@ -65,7 +65,7 @@
 ApplicationController
 =====================
 
-.. java:package:: io.github.ust.mico.core.web
+.. java:package:: io.github.ust.mico.core.resource
    :noindex:
 
 .. java:type:: @Slf4j @RestController @RequestMapping public class ApplicationController
