+++ docs/mico-core/java/io/github/ust/mico/core/web/DeploymentController.rst
@@ -57,7 +57,7 @@
 DeploymentController
 ====================
 
-.. java:package:: io.github.ust.mico.core.web
+.. java:package:: io.github.ust.mico.core.resource
    :noindex:
 
 .. java:type:: @Slf4j @RestController @RequestMapping public class DeploymentController
