+++ docs/mico-core/java/io/github/ust/mico/core/web/ServiceInterfaceController.rst
@@ -37,7 +37,7 @@
 ServiceInterfaceController
 ==========================
 
-.. java:package:: io.github.ust.mico.core.web
+.. java:package:: io.github.ust.mico.core.resource
    :noindex:
 
 .. java:type:: @Slf4j @RestController @RequestMapping public class ServiceInterfaceController
