+++ mico-core/mico-core/java/io/github/ust/mico/core/REST/ApplicationController.rst
@@ -39,7 +39,7 @@
 ApplicationController
 =====================
 
-.. java:package:: io.github.ust.mico.core.web
+.. java:package:: io.github.ust.mico.core.resource
    :noindex:
 
 .. java:type:: @RestController @RequestMapping public class ApplicationController
