+++ mico-core/mico-core/java/io/github/ust/mico/core/REST/ServiceInterfaceController.rst
@@ -45,7 +45,7 @@
 ServiceInterfaceController
 ==========================
 
-.. java:package:: io.github.ust.mico.core.web
+.. java:package:: io.github.ust.mico.core.resource
    :noindex:
 
 .. java:type:: @RestController @RequestMapping public class ServiceInterfaceController
