+++ mico-core/mico-core/java/io/github/ust/mico/core/web/ServiceController.rst
@@ -55,7 +55,7 @@
 ServiceController
 =================
 
-.. java:package:: io.github.ust.mico.core.web
+.. java:package:: io.github.ust.mico.core.resource
    :noindex:
 
 .. java:type:: @Slf4j @RestController @RequestMapping public class ServiceController
